+++ build.rs
@@ -1,0 +1,6 @@
+fn main() {
+    #[cfg(feature = "graphics")]
+    {
+        res::generate(&[res::shader("gui").transform().graphic()]);
+    }
+}